+++ src/functions.php
@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /* (c) Anton Medvedev <anton@medv.io>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -29,10 +29,9 @@ use function Deployer\Support\env_stringify;
 use function Deployer\Support\str_contains;
 
 /**
- * @param string ...$hostname
  * @return Host|Host[]|ObjectProxy
  */
-function host(...$hostname)
+function host(string ...$hostname)
 {
     $deployer = Deployer::get();
     $aliases = Range::expand($hostname);
@@ -55,7 +54,7 @@ function host(...$hostname)
         $deployer->hosts->set($aliases[0], $host);
         return $host;
     } else {
-        $hosts = array_map(function ($hostname) use ($deployer) {
+        $hosts = array_map(function ($hostname) use ($deployer) : Host {
             $host = new Host($hostname);
             $deployer->hosts->set($hostname, $host);
             return $host;
@@ -65,10 +64,9 @@ function host(...$hostname)
 }
 
 /**
- * @param string ...$hostnames
  * @return Localhost|Localhost[]|ObjectProxy
  */
-function localhost(...$hostnames)
+function localhost(string ...$hostnames)
 {
     $deployer = Deployer::get();
     $hostnames = Range::expand($hostnames);
@@ -78,7 +76,7 @@ function localhost(...$hostnames)
         $deployer->hosts->set($host->getAlias(), $host);
         return $host;
     } else {
-        $hosts = array_map(function ($hostname) use ($deployer) {
+        $hosts = array_map(function ($hostname) use ($deployer): Localhost {
             $host = new Localhost($hostname);
             $deployer->hosts->set($host->getAlias(), $host);
             return $host;
@@ -90,10 +88,8 @@ function localhost(...$hostnames)
 /**
  * Get host by host alias.
  *
- * @param string $alias
- * @return Host
  */
-function getHost(string $alias)
+function getHost(string $alias): Host
 {
     return Deployer::get()->hosts->get($alias);
 }
@@ -101,9 +97,8 @@ function getHost(string $alias)
 /**
  * Returns current host.
  *
- * @return Host
  */
-function currentHost()
+function currentHost(): Host
 {
     return Context::get()->getHost();
 }
@@ -119,7 +114,7 @@ function currentHost()
  *
  * @return Host[]
  */
-function select(string $selector)
+function select(string $selector): array
 {
     return Deployer::get()->selector->select($selector);
 }
@@ -134,15 +129,17 @@ function select(string $selector)
  * ```php
  * import(__DIR__ . '/config/hosts.yaml');
  * ```
+ *
+ * @throws Exception\Exception
  */
-function import(string $file) {
+function import(string $file): void {
     Importer::import($file);
 }
 
 /**
  * Set task description.
  */
-function desc($title = null)
+function desc(?string $title = null): ?string
 {
     static $store = null;
 
@@ -160,9 +157,9 @@ function desc($title = null)
  *
  * @param string $name Name of current task.
  * @param callable|array|string|null $body Callable task, array of other tasks names or nothing to get a defined tasks
- * @return Task
+ *
  */
-function task($name, $body = null)
+function task(string $name, $body = null): Task
 {
     $deployer = Deployer::get();
 
@@ -194,9 +191,10 @@ function task($name, $body = null)
  *
  * @param string $task The task before $that should be run.
  * @param string|callable $do The task to be run.
+ *
  * @return Task|void
  */
-function before($task, $do)
+function before(string $task, $do)
 {
     if (is_callable($do)) {
         $newTask = task("before:$task", $do);
@@ -211,9 +209,10 @@ function before($task, $do)
  *
  * @param string $task The task after $that should be run.
  * @param string|callable $do The task to be run.
+ *
  * @return Task|void
  */
-function after($task, $do)
+function after(string $task, $do)
 {
     if (is_callable($do)) {
         $newTask = task("after:$task", $do);
@@ -228,9 +227,10 @@ function after($task, $do)
  *
  * @param string $task The task which need to fail so $that should be run.
  * @param string $do The task to be run.
+ *
  * @return Task|void
  */
-function fail($task, $do)
+function fail(string $task, string $do)
 {
     if (is_callable($do)) {
         $newTask = task("fail:$task", $do);
@@ -250,7 +250,7 @@ function fail($task, $do)
  * @param string $description A description text
  * @param string|string[]|int|bool|null $default The default value (must be null for self::VALUE_NONE)
  */
-function option($name, $shortcut = null, $mode = null, $description = '', $default = null)
+function option(string $name, $shortcut = null, ?int $mode = null, string $description = '', $default = null): void
 {
     Deployer::get()->inputDefinition->addOption(
         new InputOption($name, $shortcut, $mode, $description, $default)
@@ -259,10 +259,8 @@ function option($name, $shortcut = null, $mode = null, $description = '', $defau
 
 /**
  * Change the current working directory.
- *
- * @param string $path
  */
-function cd($path)
+function cd(string $path): void
 {
     set('working_path', parse($path));
 }
@@ -270,11 +268,9 @@ function cd($path)
 /**
  * Execute a callback within a specific directory and revert back to the initial working directory.
  *
- * @param string $path
- * @param callable $callback
  * @return mixed|void Return value of the $callback function or void if callback doesn't return anything
  */
-function within($path, $callback)
+function within(string $path, callable $callback)
 {
     $lastWorkingPath = get('working_path', '');
     try {
@@ -307,13 +303,11 @@ function within($path, $callback)
  * run("echo $path");
  * ```
  *
- * @param string $command
  * @param array $options
- * @return string
  */
-function run($command, $options = [])
+function run(string $command, array $options = []): string
 {
-    $run = function ($command, $options) {
+    $run = function ($command, $options): string {
         $host = Context::get()->getHost();
 
         $command = parse($command);
@@ -372,9 +366,10 @@ function run($command, $options = [])
  *
  * @param string $command Command to run locally.
  * @param array $options
+ *
  * @return string Output of command.
  */
-function runLocally($command, $options = [])
+function runLocally(string $command, array $options = []): string
 {
     $process = Deployer::get()->processRunner;
     $command = parse($command);
@@ -399,11 +394,8 @@ function runLocally($command, $options = [])
  * ...
  * }
  * ```
- *
- * @param string $command
- * @return bool
  */
-function test($command)
+function test(string $command): bool
 {
     return run("if $command; then echo 'true'; fi") === 'true';
 }
@@ -413,11 +405,8 @@ function test($command)
  * Example:
  *
  *     testLocally('[ -d {{local_release_path}} ]')
- *
- * @param string $command
- * @return bool
  */
-function testLocally($command)
+function testLocally(string $command): bool
 {
     return runLocally("if $command; then echo 'true'; fi") === 'true';
 }
@@ -444,9 +433,8 @@ function testLocally($command)
  * ```
  *
  * @param Host|Host[] $hosts
- * @param callable $callback
  */
-function on($hosts, callable $callback)
+function on($hosts, callable $callback): void
 {
     if (!is_array($hosts) && !($hosts instanceof \Traversable)) {
         $hosts = [$hosts];
@@ -475,8 +463,10 @@ function on($hosts, callable $callback)
  * ```php
  * invoke('deploy:symlink');
  * ```
+ *
+ * @throws Exception\Exception
  */
-function invoke(string $taskName)
+function invoke(string $taskName): void
 {
     $task = Deployer::get()->tasks->get($taskName);
     Deployer::get()->messenger->startTask($task);
@@ -491,12 +481,10 @@ function invoke(string $taskName)
  * >
  * > The alternative, without the trailing slash, would place build, including the directory, within public. This would create a hierarchy that looks like: {{release_path}}/public/build
  *
- * @param string $source
- * @param string $destination
  * @param array $config
  * @throws RunException
  */
-function upload(string $source, string $destination, $config = [])
+function upload(string $source, string $destination, array $config = []): void
 {
     $rsync = Deployer::get()->rsync;
     $host = currentHost();
@@ -513,12 +501,10 @@ function upload(string $source, string $destination, $config = [])
 /**
  * Download file or directory from host
  *
- * @param string $source
- * @param string $destination
  * @param array $config
  * @throws RunException
  */
-function download(string $source, string $destination, $config = [])
+function download(string $source, string $destination, array $config = []): void
 {
     $rsync = Deployer::get()->rsync;
     $host = currentHost();
@@ -534,18 +520,16 @@ function download(string $source, string $destination, $config = [])
 
 /**
  * Writes an info message.
- * @param string $message
  */
-function info($message)
+function info(string $message): void
 {
     writeln("<fg=green;options=bold>info</> " . parse($message));
 }
 
 /**
  * Writes an warning message.
- * @param string $message
  */
-function warning($message)
+function warning(string $message): void
 {
     writeln("<fg=yellow;options=bold>warning</> <comment>" . parse($message) . "</comment>");
 }
@@ -553,9 +537,8 @@ function warning($message)
 /**
  * Writes a message to the output and adds a newline at the end.
  * @param string|array $message
- * @param int $options
  */
-function writeln($message, $options = 0)
+function writeln($message, int $options = 0): void
 {
     $host = currentHost();
     output()->writeln("[{$host->getTag()}] " . parse($message), $options);
@@ -563,11 +546,8 @@ function writeln($message, $options = 0)
 
 /**
  * Parse set values.
- *
- * @param string $value
- * @return string
  */
-function parse($value)
+function parse(string $value): string
 {
     return Context::get()->getConfig()->parse($value);
 }
@@ -575,10 +555,9 @@ function parse($value)
 /**
  * Setup configuration option.
  *
- * @param string $name
  * @param mixed $value
  */
-function set($name, $value)
+function set(string $name, $value): void
 {
     if (!Context::has()) {
         Deployer::get()->config->set($name, $value);
@@ -590,10 +569,9 @@ function set($name, $value)
 /**
  * Merge new config params to existing config array.
  *
- * @param string $name
  * @param array $array
  */
-function add($name, $array)
+function add(string $name, array $array): void
 {
     if (!Context::has()) {
         Deployer::get()->config->add($name, $array);
@@ -605,11 +583,11 @@ function add($name, $array)
 /**
  * Get configuration value.
  *
- * @param string $name
  * @param mixed|null $default
+ *
  * @return mixed
  */
-function get($name, $default = null)
+function get(string $name, $default = null)
 {
     if (!Context::has()) {
         return Deployer::get()->config->get($name, $default);
@@ -620,11 +598,8 @@ function get($name, $default = null)
 
 /**
  * Check if there is such configuration option.
- *
- * @param string $name
- * @return boolean
  */
-function has($name)
+function has(string $name): bool
 {
     if (!Context::has()) {
         return Deployer::get()->config->has($name);
@@ -634,12 +609,9 @@ function has($name)
 }
 
 /**
- * @param string $message
- * @param string|null $default
  * @param string[]|null $autocomplete
- * @return string
  */
-function ask($message, $default = null, $autocomplete = null)
+function ask(string $message, ?string $default = null, ?array $autocomplete = null): ?string
 {
     Context::required(__FUNCTION__);
 
@@ -666,13 +638,12 @@ function ask($message, $default = null, $autocomplete = null)
 }
 
 /**
- * @param string $message
  * @param string[] $availableChoices
- * @param string|null $default
  * @param bool|false $multiselect
+ *
  * @return string|string[]
  */
-function askChoice($message, array $availableChoices, $default = null, $multiselect = false)
+function askChoice(string $message, array $availableChoices, ?string $default = null, bool $multiselect = false)
 {
     Context::required(__FUNCTION__);
 
@@ -706,12 +677,7 @@ function askChoice($message, array $availableChoices, $default = null, $multisel
     return $helper->ask(input(), output(), $question);
 }
 
-/**
- * @param string $message
- * @param bool $default
- * @return bool
- */
-function askConfirmation($message, $default = false)
+function askConfirmation(string $message, bool $default = false): bool
 {
     Context::required(__FUNCTION__);
 
@@ -734,11 +700,7 @@ function askConfirmation($message, $default = false)
     return $helper->ask(input(), output(), $question);
 }
 
-/**
- * @param string $message
- * @return string
- */
-function askHiddenResponse(string $message)
+function askHiddenResponse(string $message): string
 {
     Context::required(__FUNCTION__);
 
@@ -762,35 +724,25 @@ function askHiddenResponse(string $message)
     return $helper->ask(input(), output(), $question);
 }
 
-/**
- * @return InputInterface
- */
-function input()
+function input(): InputInterface
 {
     return Context::get()->getInput();
 }
 
-
-/**
- * @return OutputInterface
- */
-function output()
+function output(): OutputInterface
 {
     return Context::get()->getOutput();
 }
 
 /**
  * Check if command exists
- *
- * @param string $command
- * @return bool
  */
-function commandExist($command)
+function commandExist(string $command): bool
 {
     return test("hash $command 2>/dev/null");
 }
 
-function commandSupportsOption($command, $option)
+function commandSupportsOption(string $command, string $option): bool
 {
     $man = run("(man $command 2>&1 || $command -h 2>&1 || $command --help 2>&1) | grep -- $option || true");
     if (empty($man)) {
@@ -799,7 +751,7 @@ function commandSupportsOption($command, $option)
     return str_contains($man, $option);
 }
 
-function locateBinaryPath($name)
+function locateBinaryPath(string $name): string
 {
     $nameEscaped = escapeshellarg($name);