+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Added support for placeholders in `run()` func.
 - Added support for secret passing in `run()` func without outputting to logs.
 - Added docker-based E2E testing environment. [#2197]
+- Typehinting for functions.php. [#2187]
 
 ### Changed
 - Refactored executor engine, up to 2x faster than before.
@@ -598,6 +599,7 @@
 
 
 [#2197]: https://github.com/deployphp/deployer/issues/2197
+[#2187]: https://github.com/deployphp/deployer/issues/2187
 [#1994]: https://github.com/deployphp/deployer/issues/1994
 [#1990]: https://github.com/deployphp/deployer/issues/1990
 [#1989]: https://github.com/deployphp/deployer/issues/1989